Structure the code
