--- README.txt
@@ -1,1 +1,0 @@
-'informatie over installatie'
