+++ src/data/old/icons.qrc
@@ -1,0 +1,5 @@
+<!DOCTYPE RCC><RCC version="1.0">
+<qresource>
+    <file>/images/icons/monitor.png</file>
+</qresource>
+</RCC>
