+++ src/data/test_data/should_work.csv
@@ -1,0 +1,2 @@
+Niels,1234,Niels van den Bos,Zelfstandige,Ja
+Jeroen,ergogjnerofn,Jeroen Brauns,Heijmans Safety & Security, Ja
