+++ src/icons/monitor-icon.png
