--- wdt/old/icons.qrc
@@ -1,5 +1,0 @@
-<!DOCTYPE RCC><RCC version="1.0">
-<qresource>
-    <file>/images/icons/monitor.png</file>
-</qresource>
-</RCC>
